+++ src/ell_stable.rs
@@ -59,7 +59,7 @@ impl EllStable {
     /// an instance of the [`EllStable`] struct.
     pub fn new_with_matrix(kappa: f64, mq: Array2<f64>, xc: Array1<f64>) -> EllStable {
         let ndim = xc.len();
-        let helper = EllCalc::new(ndim as f64);
+        let helper = EllCalc::new(ndim);
 
         EllStable {
             kappa,
